+++ app/src/js/components/NetworkErrorModal/network-error-modal.js
@@ -1,0 +1,47 @@
+import React, { useEffect, useState } from 'react';
+import PropTypes from 'prop-types';
+import DefaultModal from '../Modal/modal';
+
+const NetworkErrorModal = ({
+  title = 'Network Connection Error',
+  children = 'Your network appears to be offline. Please check your internet connection and refresh the page.',
+}) => {
+  const [modalVisible, setModalVisible] = useState(!navigator.onLine);
+  const handleClose = () => setModalVisible(false);
+  useEffect(() => {
+    const updateStatus = () => {
+      const offline = !navigator.onLine;
+      setModalVisible(offline);
+    };
+    window.addEventListener('online', updateStatus);
+    window.addEventListener('offline', updateStatus);
+    const interval = setInterval(updateStatus, 1000); // check every second
+
+    return () => {
+      window.removeEventListener('online', updateStatus);
+      window.removeEventListener('offline', updateStatus);
+      clearInterval(interval);
+    };
+  }, []);
+
+  return (
+    <DefaultModal
+    title={title}
+    className="NetworkErrorModal"
+    onCancel={handleClose}
+    onCloseModal={handleClose}
+    showModal={modalVisible}
+    hasConfirmButton={false}
+    hasCancelButton={true}
+    cancelButtonText="Dismiss"
+    >
+        {children}
+        </DefaultModal>
+  );
+};
+NetworkErrorModal.propTypes = {
+  title: PropTypes.string,
+  children: PropTypes.node,
+};
+
+export default NetworkErrorModal;